+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `script vivado`: Add options to restrict output to defines, includes, or sources.
+
 ### Changed
 - `script`: Raise error if `vcom-arg` or `vlog-arg` are used with format other than `vsim`.
 - `script vivado`: